+++ src/modules/api/spoonacular/spoonacular.api.types.ts
@@ -3,7 +3,7 @@ import { IngredientName, IngredientUnit } from '../../../common/enums';
 export type SpoonacularIngredient = {
     amount: number,
     id: number,
-    imageUrl: string,
+    image: string,
     meta: [],
     name: IngredientName,
     unit: IngredientUnit
@@ -24,5 +24,26 @@ export type SpoonacularRecipeDetails = {
     id: number,
     image: string,
     title: string,
-    extendedIngredients: SpoonacularIngredient[]
+    extendedIngredients: SpoonacularIngredient[],
+    summary: string,
+    readyInMinutes: number;
+    vegetarian: boolean;
+    vegan: boolean;
+    glutenFree: boolean;
+    dairyFree: boolean;
+    veryHealthy: boolean;
+    creditsText: string;
+    sourceName: string;
 };
+
+type SpoonacularMealInstructionStep = {
+    number: number,
+    step: string
+};
+
+export type SpoonacularMealInstructionRecord = {
+    name: string,
+    steps: SpoonacularMealInstructionStep[]
+};
+
+export type SpoonacularMealInstruction = SpoonacularMealInstructionRecord[];