+++ src/mongodb/documents/meal.document.ts
@@ -1,5 +1,6 @@
 import { Document } from 'mongoose';
 import { MealIngredient } from '../../modules/ingredient/ingredient.types';
+import { MealInstruction } from '../../modules/meal/meal.types';
 
 export interface MealDocument extends Document {
   readonly author: string;
@@ -12,4 +13,6 @@ export interface MealDocument extends Document {
   readonly softAdded?: boolean;
   readonly softDeleted?: boolean;
   readonly softEdited?: MealDocument;
+  readonly readyInMinutes: number;
+  readonly instruction: MealInstruction;
 }