+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.31.0] - 2024-07-10
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): More properties of meal, e.g. is vegan, healthy and so on, and how much time it takes to prepare.
+- [Marek Kurańda](https://github.com/mjkuranda): Instruction info from Spoonacular API.
+
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): Ingredient and meal images URL.
+- [Marek Kurańda](https://github.com/mjkuranda): API url in abstract service.
+
 ## [1.30.0] - 2024-07-10
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Enabled CORS for `http://localhost:3000/`.